feat: New config option: data-host
If set to a non-empty string, the data-host config option will force all HTTP
traffic sent by the `DataProcessor` through the specified host. When
`DataProcessor` is processing a message, if the worker has `detachedContent`
set to `true`, the value of the message's `content` field is parsed as a URL
and the value of the URL's `Host` field is set to the value specified by the
data-host argument before sending the HTTP request to the HTTP client for
handling.

If no value is set in either the config file or on the command line, a
compile-time constant, `DataHost` is set as the default value. In any case, if
the value of `DataHost` is an empty string, no host substitution is performed.

Fixes RHCLOUD-12548