fix: Change message_router routine startup logic
The message_router routine startup logic now connects the MQTT client *and* subscribes to topics. It blocks publishing a connection-status message until it receives a value on the process-bootstrap signal. Only after receiving this signal and publishing a connection-status message will it connect to the worker-register signals. This should prevent the flurry of connection-status messages from occurring at startup.